+++ src/pages/user/EditInfoPage.tsx
@@ -1,8 +1,5 @@
 import { ChangeEvent, useRef, useState } from 'react';
-import FilterList from '../../components/Common/FilterList';
-import TitleIntro from '../../components/Common/TitleIntro';
-import ProfileImg from '../../components/Common/ProfileImg';
-import { useFilterListStore, useUserStore } from '@/config/store';
+import TitleIntro from '@/components/Common/TitleIntro';
 import Nickname from '@/components/EditInfoPage/Nickname';
 import Description from '@/components/EditInfoPage/Description';
 import GenderCheck from '@/components/EditInfoPage/GenderCheck';
@@ -11,6 +8,8 @@ import { updateMyProfile } from '@/api/user';
 import { UserProfileUpdateData } from '@/config/types';
 import { toast, ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import MypageLayout from '@/layouts/MypageLayout';
+import { useUserStore } from '@/config/store';
 
 export default function EditInfoPage() {
   const { user, setUser } = useUserStore();
@@ -35,9 +34,6 @@ export default function EditInfoPage() {
     gender: user.gender,
     birthday: birthDate === null ? null : `${birthYear}-${birthMonth}-${birthDay}`,
   });
-
-  const isFilterListOpen = useFilterListStore((state) => state.isFilterListOpen);
-  const setIsFilterListToggle = useFilterListStore((state) => state.setIsFilterListToggle);
 
   const handleChangePickedImage = (e: ChangeEvent<HTMLInputElement>) => {
     const file = e.target.files![0];
@@ -99,73 +95,52 @@ export default function EditInfoPage() {
   return (
     <>
       <ToastContainer />
-      <div className='h-full w-full'>
-        <TitleIntro titleE={'MY PROFILE'} titleK={'프로필 편집'} content={'멋진 실력을 자랑해주세요!'} />
-        <div className='flex h-full w-full transition-all duration-300 ease-in-out'>
-          {/* 필터 섹션 */}
-          <div
-            className={`relative z-30 transition-all duration-300 ease-in-out ${
-              isFilterListOpen ? 'max-w-[300px] lg:max-w-[350px] xl:max-w-[400px]' : 'max-w-0'
-            }`}
-          >
-            <FilterList isFilterListOpen={isFilterListOpen} setIsFilterListToggle={setIsFilterListToggle} />
-          </div>
-          {/* 메인 섹션 */}
-          <div className={`relative flex-grow bg-gray-100 p-20 transition-all duration-300 ease-in-out`}>
-            <ProfileImg />
-            {/* 이벤트 배너 */}
-            <div className='m-auto mt-[140px] max-w-[800px]'>
-              <div className='w-full bg-gray-100'>
-                {previewImage ? (
-                  <div>
-                    <img
-                      className='w-min-[100px] h-[300px] w-[52.6%] rounded-full object-cover'
-                      src={previewImage}
-                      alt='이미지 미리보기'
-                    />
-                  </div>
-                ) : null}
-                <form onSubmit={handleSubmit} className='flex w-full flex-col items-end'>
-                  <label
-                    htmlFor='fileInput'
-                    className='inline-block transform cursor-pointer rounded-xl bg-primary px-3 py-2 text-[24px] font-bold hover:scale-95'
-                  >
-                    프로필 사진 올리기
-                    <input
-                      id='fileInput'
-                      ref={fileRef}
-                      onChange={handleChangePickedImage}
-                      type='file'
-                      className='hidden'
-                    />
-                  </label>
-
-                  <Nickname profile={profile} setProfile={setProfile} />
-                  <Description profile={profile} setProfile={setProfile} />
-                  <GenderCheck profile={profile} setProfile={setProfile} />
-                  <Birthday
-                    profile={profile}
-                    setProfile={setProfile}
-                    birthYear={birthYear}
-                    setBirthYear={setBirthYear}
-                    birthMonth={birthMonth}
-                    setBirthMonth={setBirthMonth}
-                    birthDay={birthDay}
-                    setBirthDay={setBirthDay}
-                  />
-
-                  <button
-                    className='h-[66px] w-full rounded-xl bg-primary text-2xl font-bold text-gray-500 hover:scale-95'
-                    type='submit'
-                  >
-                    저장
-                  </button>
-                </form>
+      <TitleIntro titleE={'MY PROFILE'} titleK={'프로필 편집'} content={'멋진 실력을 자랑해주세요!'} />
+      <MypageLayout>
+        <div className='m-auto mt-[140px] max-w-[800px]'>
+          <div className='w-full bg-gray-100'>
+            {previewImage ? (
+              <div>
+                <img
+                  className='w-min-[100px] h-[300px] w-[52.6%] rounded-full object-cover'
+                  src={previewImage}
+                  alt='이미지 미리보기'
+                />
               </div>
-            </div>
+            ) : null}
+            <form onSubmit={handleSubmit} className='flex w-full flex-col items-end'>
+              <label
+                htmlFor='fileInput'
+                className='inline-block transform cursor-pointer rounded-xl bg-primary px-3 py-2 text-[24px] font-bold hover:scale-95'
+              >
+                프로필 사진 올리기
+                <input id='fileInput' ref={fileRef} onChange={handleChangePickedImage} type='file' className='hidden' />
+              </label>
+
+              <Nickname profile={profile} setProfile={setProfile} />
+              <Description profile={profile} setProfile={setProfile} />
+              <GenderCheck profile={profile} setProfile={setProfile} />
+              <Birthday
+                profile={profile}
+                setProfile={setProfile}
+                birthYear={birthYear}
+                setBirthYear={setBirthYear}
+                birthMonth={birthMonth}
+                setBirthMonth={setBirthMonth}
+                birthDay={birthDay}
+                setBirthDay={setBirthDay}
+              />
+
+              <button
+                className='h-[66px] w-full rounded-xl bg-primary text-2xl font-bold text-gray-500 hover:scale-95'
+                type='submit'
+              >
+                저장
+              </button>
+            </form>
           </div>
         </div>
-      </div>
+      </MypageLayout>
     </>
   );
 }
